requirements.txt
python-crontab==2.6.0
requests==2.27.1